backbanco/projeto_bradescov2/projeto_bradescov2/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
PyJWT